requirements-main.txt
torch-ac>=1.0.1
gym-minigrid
tensorboardX>=1.6
torch>=1
networkx>=2.1
numpy>=1.3